op3(t): tame qti-telephony-common
after spending multiple hours with logs this solves
the problem of correct save and restore of default cards
assignments in the msim case.

data:
set that hidden setting and the iccid in the two
properties make the selection persist accross reboots
it would also be possible to the the iccid to 0
then they will be reset on next reboot. Unknown
if that would have other side-effects

calls:
setting property persist.sys.phone_account will persist
the active value for calls. If the property
persist.radio.aosp_usr_pref_sel=true is set this would
not be required. But also here its unknowne if that
would have any other side-effects

Change-Id: Iff8f420ae570ebcaf162ae20455911ba36aef31b